requirements.txt
pytube 
youtube-transcript-api

google-genai
python-dotenv
google-api-python-client

curl_cffi 